+++ apps/journeys/src/components/Conductor/Conductor.spec.tsx
@@ -2,7 +2,7 @@ import { useBreakpoints } from '@core/shared/ui'
 import Conductor from '.'
 import { fireEvent, renderWithApolloClient } from '../../../test/testingLibrary'
 import { activeBlockVar, treeBlocksVar } from '../../libs/client/cache/blocks'
-import { TreeBlock } from '../../libs/transformer/transformer'
+import { TreeBlock } from '../../libs/TreeBlock'
 
 jest.mock('../../../../../libs/shared/ui/src/', () => ({
   __esModule: true,
