+++ apps/journeys/src/components/JourneyProgress/JourneyProgress.stories.tsx
@@ -2,7 +2,7 @@ import { journeysConfig, simpleComponentConfig } from '../../libs/storybook'
 import { Meta } from '@storybook/react'
 import { ReactElement } from 'react'
 import { JourneyProgress } from '.'
-import { TreeBlock } from '../../libs/transformer/transformer'
+import { TreeBlock } from '../../libs/TreeBlock'
 import { activeBlockVar, treeBlocksVar } from '../../libs/client/cache/blocks'
 import { GetJourney_journey_blocks_StepBlock as StepBlock } from '../../../__generated__/GetJourney'
