+++ apps/journeys/src/components/blocks/Card/Card.spec.tsx
@@ -1,6 +1,6 @@
 import { render } from '../../../../test/testingLibrary'
 import { Card } from '.'
-import { TreeBlock } from '../../../libs/transformer/transformer'
+import { TreeBlock } from '../../../libs/TreeBlock'
 import { ThemeMode, ThemeName } from '../../../../__generated__/globalTypes'
 import { themes } from '@core/shared/ui'
