+++ apps/journeys/src/components/blocks/GridContainer/GridContainer.stories.tsx
@@ -1,6 +1,6 @@
 import { Story, Meta } from '@storybook/react'
 import { GetJourney_journey_blocks_GridContainerBlock as GridContainerBlock } from '../../../../__generated__/GetJourney'
-import { TreeBlock } from '../../../libs/transformer/transformer'
+import { TreeBlock } from '../../../libs/TreeBlock'
 import { GridContainer } from './GridContainer'
 import { journeysConfig, StoryCard } from '../../../libs/storybook'
 import {
