+++ apps/journeys/src/components/blocks/GridItem/GridItem.stories.tsx
@@ -1,6 +1,6 @@
 import { Story, Meta } from '@storybook/react'
 import { GetJourney_journey_blocks_GridItemBlock as GridItemBlock } from '../../../../__generated__/GetJourney'
-import { TreeBlock } from '../../../libs/transformer/transformer'
+import { TreeBlock } from '../../../libs/TreeBlock'
 import { GridItem } from './GridItem'
 import { journeysConfig, StoryCard } from '../../../libs/storybook'
 import {
