+++ apps/journeys/src/components/blocks/Image/Image.spec.tsx
@@ -1,7 +1,7 @@
 import { render } from '../../../../test/testingLibrary'
 import { Image } from '.'
 import { GetJourney_journey_blocks_ImageBlock as ImageBlock } from '../../../../__generated__/GetJourney'
-import { TreeBlock } from '../../../libs/transformer/transformer'
+import { TreeBlock } from '../../../libs/TreeBlock'
 
 describe('Image', () => {
   const block: TreeBlock<ImageBlock> = {
