+++ apps/journeys/src/components/blocks/Image/Image.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react'
 import NextImage from 'next/image'
 import { GetJourney_journey_blocks_ImageBlock as ImageBlock } from '../../../../__generated__/GetJourney'
-import { TreeBlock } from '../../../libs/transformer/transformer'
+import { TreeBlock } from '../../../libs/TreeBlock'
 import { SxProps } from '@mui/system'
 import { Theme, Box } from '@mui/material'
