+++ apps/journeys/src/components/blocks/RadioQuestion/RadioQuestion.tsx
@@ -2,7 +2,7 @@ import { ReactElement } from 'react'
 import { Typography, ButtonGroup } from '@mui/material'
 import { RadioOption } from '../RadioOption'
 import { GetJourney_journey_blocks_RadioQuestionBlock as RadioQuestionBlock } from '../../../../__generated__/GetJourney'
-import { TreeBlock } from '../../../libs/transformer/transformer'
+import { TreeBlock } from '../../../libs/TreeBlock'
 import { useMutation, gql } from '@apollo/client'
 import { RadioQuestionResponseCreate } from '../../../../__generated__/RadioQuestionResponseCreate'
 import { v4 as uuidv4 } from 'uuid'
