+++ apps/journeys/src/components/blocks/Step/Step.tsx
@@ -1,4 +1,4 @@
-import { TreeBlock } from '../../../libs/transformer/transformer'
+import { TreeBlock } from '../../../libs/TreeBlock'
 import { ReactElement } from 'react'
 import { BlockRenderer } from '../../BlockRenderer'
 import { GetJourney_journey_blocks_StepBlock as StepBlock } from '../../../../__generated__/GetJourney'
