+++ apps/journeys/src/components/blocks/Typography/Typography.stories.tsx
@@ -15,7 +15,7 @@ import {
   StoryCard
 } from '../../../libs/storybook'
 import { GetJourney_journey_blocks_TypographyBlock as TypographyBlock } from '../../../../__generated__/GetJourney'
-import { TreeBlock } from '../../../libs/transformer/transformer'
+import { TreeBlock } from '../../../libs/TreeBlock'
 
 const TypographyDemo = {
   ...journeysConfig,
