+++ apps/journeys/src/libs/TreeBlock/TreeBlock.ts
@@ -1,0 +1,4 @@
+import { GetJourney_journey_blocks as Block } from '../../../__generated__/GetJourney'
+import { TreeBlock as CoreTreeBlock } from '@core/shared/ui'
+
+export type TreeBlock<ChildBlock = Block> = CoreTreeBlock<ChildBlock, Block>
