+++ apps/journeys/src/libs/TreeBlock/index.ts
@@ -1,0 +1,1 @@
+export { TreeBlock } from './TreeBlock'
