+++ apps/journeys/src/libs/client/cache/blocks.ts
@@ -1,7 +1,7 @@
 import { makeVar, useReactiveVar } from '@apollo/client'
 import { GetJourney_journey_blocks_StepBlock as StepBlock } from '../../../../__generated__/GetJourney'
 import { useCallback } from 'react'
-import { TreeBlock } from '../../transformer/transformer'
+import { TreeBlock } from '../../TreeBlock'
 
 export const activeBlockVar = makeVar<TreeBlock<StepBlock> | null>(null)
 export const previousBlocksVar = makeVar<TreeBlock[]>([])
