--- apps/journeys/src/libs/transformer/index.ts
@@ -1,3 +1,0 @@
-import transformer from './transformer'
-
-export default transformer
