+++ libs/shared/ui/src/index.ts
@@ -2,3 +2,4 @@ export { ThemeProvider } from './components/ThemeProvider'
 export { sharedUiConfig } from './libs/storybook/config'
 export { themes } from './libs/themes'
 export { useBreakpoints } from './libs/useBreakpoints'
+export { transformer, TreeBlock } from './libs/transformer'
