+++ libs/shared/ui/src/libs/transformer/index.ts
@@ -1,0 +1,1 @@
+export { transformer, TreeBlock } from './transformer'
